+++ app/src/main/java/com/swent/assos/ui/screens/ticket/MyTickets.kt
@@ -3,25 +3,40 @@
 package com.swent.assos.ui.screens.ticket
 
 import android.net.Uri
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.semantics.testTagsAsResourceId
+import androidx.compose.ui.unit.dp
+import coil.compose.rememberAsyncImagePainter
 import com.swent.assos.model.data.Ticket
+import com.swent.assos.model.navigation.Destinations
 import com.swent.assos.model.navigation.NavigationActions
 import com.swent.assos.ui.components.PageTitle
-import com.swent.assos.ui.components.TicketItem
 import java.time.LocalDateTime
 
 @OptIn(ExperimentalComposeUiApi::class)
@@ -49,7 +64,7 @@ fun MyTickets(navigationActions: NavigationActions) {
   val myTickets = listOf(balelecTicket, rocketTeamTicket, challengeTicket, amacPartyTicket)
 
   Scaffold(
-      modifier = Modifier.semantics { testTagsAsResourceId = true }.testTag("TicketScreen"),
+      modifier = Modifier.semantics { testTagsAsResourceId = true }.testTag("MyTicketsScreen"),
       topBar = { PageTitle(title = "My Tickets") }) { paddingValues ->
         LazyColumn(
             modifier = Modifier.fillMaxWidth().padding(paddingValues).testTag("TicketList"),
@@ -61,4 +76,52 @@ fun MyTickets(navigationActions: NavigationActions) {
           }
         }
       }
+}
+
+@Composable
+fun TicketItem(ticket: Ticket, navigationActions: NavigationActions) {
+    Card(
+        colors = CardDefaults.cardColors(MaterialTheme.colorScheme.onPrimary),
+        shape = RoundedCornerShape(12.dp),
+        modifier =
+        Modifier.testTag("TicketItem")
+            .padding(16.dp)
+            .border(
+                width = 0.5.dp,
+                color = MaterialTheme.colorScheme.outline,
+                shape = RoundedCornerShape(12.dp)
+            )) {
+        Column(
+            modifier =
+            Modifier.fillMaxWidth().padding(vertical = 0.dp).clickable {
+                navigationActions.navigateTo(Destinations.TICKET_DETAILS.route)
+            },
+        ) {
+            Image(
+                painter = rememberAsyncImagePainter(ticket.banner),
+                contentDescription = null,
+                contentScale = ContentScale.Crop,
+                modifier =
+                Modifier.fillMaxWidth()
+                    .height(84.dp)
+                    .background(MaterialTheme.colorScheme.outline))
+
+            Spacer(modifier = Modifier.height(10.dp))
+            Text(
+                text = ticket.name,
+                style = MaterialTheme.typography.titleMedium,
+                modifier = Modifier.fillMaxWidth().padding(horizontal = 16.dp))
+            Spacer(modifier = Modifier.height(6.dp))
+
+            Text(
+                text = ticket.startTime?.let { dateToReadableString(it) } ?: "",
+                style = MaterialTheme.typography.bodyMedium,
+                modifier = Modifier.fillMaxWidth().padding(horizontal = 16.dp))
+            Spacer(modifier = Modifier.height(10.dp))
+        }
+    }
+}
+
+fun dateToReadableString(date: LocalDateTime): String {
+    return "${date.dayOfMonth} ${date.month} ${date.year}, ${date.hour}:${date.minute}"
 }
